fastapi_server/requirements.txt
# fastapi_server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
slowapi
pydantic
